+++ include/standardese/cpp_cursor.hpp
@@ -5,9 +5,12 @@
 #ifndef STANDARDESE_CPP_CURSOR_HPP_INCLUDED
 #define STANDARDESE_CPP_CURSOR_HPP_INCLUDED
 
+#include <cassert>
+
 #include <clang-c/Index.h>
 
 #include <standardese/noexcept.hpp>
+#include <standardese/string.hpp>
 
 namespace standardese
 {
@@ -26,7 +29,7 @@ namespace standardese
 
     inline bool operator==(const cpp_cursor& a, const cpp_cursor& b) STANDARDESE_NOEXCEPT
     {
-        return clang_equalCursors(a, b) == 1u;
+        return clang_equalCursors(a, b) != 0u;
     }
 
     inline bool operator!=(const cpp_cursor& a, const cpp_cursor& b) STANDARDESE_NOEXCEPT
@@ -34,17 +37,5 @@ namespace standardese
         return !(a == b);
     }
 } // namespace standardese
-
-namespace std
-{
-    template <>
-    struct hash<standardese::cpp_cursor>
-    {
-        std::size_t operator()(const standardese::cpp_cursor& cur) const STANDARDESE_NOEXCEPT
-        {
-            return clang_hashCursor(cur);
-        }
-    };
-} // namespace std
 
 #endif //STANDARDESE_CPP_CURSOR_HPP_INCLUDED