+++ src/comment.cpp
@@ -111,14 +111,11 @@ namespace
                 escape = false;
                 result += std::string(tab_width, ' ');
             }
-            else if (escape)
-            {
-                result += '\\';
-                result += c;
-                escape = false;
-            }
             else
+            {
+                escape = false;
                 result += c;
+            }
     }
 }
 
@@ -194,7 +191,8 @@ string detail::get_unique_name(const doc_entity* parent, const string& unique_na
     std::string result;
     if (need_parent_name(parent))
     {
-        if (parent->get_entity_type() == doc_entity::member_group_t && parent->has_parent())
+        if (parent->get_entity_type() == doc_entity::member_group_t
+            && need_parent_name(parent->has_parent() ? &parent->get_parent() : nullptr))
             // use parent of group
             result += parent->get_parent().get_unique_name().c_str();
         else if (parent->get_entity_type() != doc_entity::member_group_t)
@@ -825,6 +823,8 @@ namespace
                 while (*end_id && !std::isspace(*end_id))
                     ++end_id;
                 stack.info().comment.add_to_member_group(get_group_id(arg, end_id));
+                while (*end_id && std::isspace(*end_id))
+                    ++end_id;
                 if (*end_id)
                     stack.info().comment.set_group_name(end_id);
                 break;